+++ favoriteCharacter.html
@@ -5,17 +5,45 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="shortcut icon" href="assets/images/logo.png" />
-    <title>Lord Of The Rings - Favoriete karakter</title>
+    <title>Lord Of The Rings - Gandalf</title>
     <link rel="stylesheet" href="css/reset.css">
     <link rel="stylesheet" href="css/style.css">
+    <script type="module" src="js/characterDetail.js" defer></script>
 </head>
 
 <body class="favCharacter">
     <header>
+        <figure>
+            <a href="index.html">
+                <img src="assets/images/logo.png" alt="logo Lord Of The Rings" height="100px">
+            </a>
+        </figure>
 
+        <nav>
+            <ul>
+                <li><a href="login.html">Login</a></li>
+                <li><a href="registration.html">Register</a></li>
+                <li><a href="index.html">Home</a></li>
+                <li><a href="favorites.html">Favorieten</a></li>
+                <li><a href="blacklist.html">Zwarte lijst</a></li>
+            </ul>
+        </nav>
     </header>
     <main>
+        <h2>Info over ...</h2>
+        <section class="profile">
+            <h3>Gandalf</h3>
+            <figure>
+                <img src="assets/images/character_Gandalf.jpeg" alt="Foto van Gandalf" height="300px">
+            </figure>
+        </section>
+        <secition class="info">
+            <h3>Informatie over Gandalf</h3>
 
+        </secition>
+        <section class="quotes">
+            <h3>Quotes</h3>
+        </section>
     </main>
     <footer>
         &copy; Copyright WPL Projects 2024-2025 - Abdul Kosimov, Hakki Turaç, Metin Molov, Nabil Habrani
