+++ favorites.html
@@ -8,16 +8,25 @@
     <title>Lord Of The Rings - Favorieten</title>
     <link rel="stylesheet" href="css/reset.css">
     <link rel="stylesheet" href="css/style.css">
-    <script type="module" src="js/api.js" defer></script>
+    <script type="module" src="js/listAllCharacters.js" defer></script>
 </head>
 
 <body class="favorites">
     <header>
         <figure>
-            <a href="#">
+            <a href="index.html">
                 <img src="assets/images/logo.png" alt="logo Lord Of The Rings" height="100px">
             </a>
         </figure>
+        <nav>
+            <ul>
+                <li><a href="login.html">Login</a></li>
+                <li><a href="registration.html">Register</a></li>
+                <li><a href="index.html">Home</a></li>
+                <li><a href="favorites.html">Favorieten</a></li>
+                <li><a href="blacklist.html">Zwarte lijst</a></li>
+            </ul>
+        </nav>
     </header>
 
     <h2>Je favoriete karakters</h2>
