+++ registration.html
@@ -29,14 +29,14 @@
 
             <label for="password">Wachtwoord</label>
             <input type="password" id="password" name="password" placeholder="*****">
-            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="toggleVisibillityOnMouseDown(event)"
-                onmouseup="toggleVisibillityOnMouseUp(event)"></i>
+            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="togglePasswordOnMouseDown(event)"
+                onmouseup="togglePasswordOnMouseUp(event)"></i>
             <br>
 
             <label for="confirm-password">Bevestig uw wachtwoord</label>
             <input type="password" id="confirm-password" name="confirm-password" placeholder="*****">
-            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="toggleVisibillityOnMouseDown(event)"
-                onmouseup="toggleVisibillityOnMouseUp(event)"></i>
+            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="togglePasswordOnMouseDown(event)"
+                onmouseup="togglePasswordOnMouseUp(event)"></i>
             <br>
 
             <a href="resetPassword.html">
