+++ resetPassword.html
@@ -21,19 +21,18 @@
         </figure>
     </header>
     <main>
-        <h2>Herstel wachtwoord</h2>
+        <h2>Herstel je wachtwoord</h2>
         <form method="post">
             <label for="resetPassword1">Geef nieuwe wachtwoord in:</label>
             <input type="password" name="resetPassword1" id="resetPassword1" placeholder="*****">
-            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="toggleVisibillityOnMouseDown(event)"
-                onmouseup="toggleVisibillityOnMouseUp(event)"></i>
+            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="togglePasswordOnMouseDown(event)"
+                onmouseup="togglePasswordOnMouseUp(event)"></i>
             <br>
-
 
             <label for="resetPassword2">Bevestig nieuwe wachtwoord:</label>
             <input type="password" name="resetPassword2" id="resetPassword2" placeholder="*****">
-            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="toggleVisibillityOnMouseDown(event)"
-                onmouseup="toggleVisibillityOnMouseUp(event)"></i>
+            <i class="fa fa-eye eyeIcon eyeIcon1" onmousedown="togglePasswordOnMouseDown(event)"
+                onmouseup="togglePasswordOnMouseUp(event)"></i>
 
             <button type="submit" id="submit">Herstel wachtwoord</button>
         </form>
